Factorize common UI painting code to ui/internal/paintutil.h/cpp (#252)
