+++ libs/vgc/ui/button.cpp
@@ -17,11 +17,10 @@
 #include <vgc/ui/button.h>
 
 #include <vgc/core/floatarray.h>
-#include <vgc/core/paths.h>
-#include <vgc/core/random.h>
-#include <vgc/graphics/font.h>
 #include <vgc/ui/strings.h>
 #include <vgc/ui/style.h>
+
+#include <vgc/ui/internal/paintutil.h>
 
 namespace vgc {
 namespace ui {
@@ -65,158 +64,19 @@ void Button::onPaintCreate(graphics::Engine* engine)
     trianglesId_ = engine->createTriangles();
 }
 
-namespace {
-
-void insertTriangle(
-        core::FloatArray& a,
-        float r, float g, float b,
-        float x1, float y1,
-        float x2, float y2,
-        float x3, float y3)
-{
-    a.insert(a.end(), {
-        x1, y1, r, g, b,
-        x2, y2, r, g, b,
-        x3, y3, r, g, b});
-}
-
-void insertRect(
-        core::FloatArray& a,
-        float r, float g, float b,
-        float x1, float y1, float x2, float y2)
-{
-    a.insert(a.end(), {
-        x1, y1, r, g, b,
-        x2, y1, r, g, b,
-        x1, y2, r, g, b,
-        x2, y1, r, g, b,
-        x2, y2, r, g, b,
-        x1, y2, r, g, b});
-}
-
-void insertRect(
-        core::FloatArray& a,
-        const core::Color& c,
-        float x1, float y1, float x2, float y2,
-        float borderRadius)
-{
-    float r = static_cast<float>(c[0]);
-    float g = static_cast<float>(c[1]);
-    float b = static_cast<float>(c[2]);
-    float maxBorderRadius = 0.5 * std::min(std::abs(x2-x1), std::abs(y2-y1));
-    borderRadius = core::clamp(borderRadius, 0.0f, maxBorderRadius);
-    Int32 numCornerTriangles = core::ifloor<Int32>(borderRadius);
-    if (numCornerTriangles < 1) {
-        a.insert(a.end(), {
-            x1, y1, r, g, b,
-            x2, y1, r, g, b,
-            x1, y2, r, g, b,
-            x2, y1, r, g, b,
-            x2, y2, r, g, b,
-            x1, y2, r, g, b});
-    }
-    else {
-        float x1_ = x1 + borderRadius;
-        float x2_ = x2 - borderRadius;
-        float y1_ = y1 + borderRadius;
-        float y2_ = y2 - borderRadius;
-        // center rectangle
-        insertRect(a, r, g, b, x1_, y1_, x2_, y2_);
-        // side rectangles
-        insertRect(a, r, g, b, x1_, y1, x2_, y1_);
-        insertRect(a, r, g, b, x2_, y1_, x2, y2_);
-        insertRect(a, r, g, b, x1_, y2_, x2_, y2);
-        insertRect(a, r, g, b, x1, y1_, x1_, y2_);
-        // rounded corners
-        float rcost_ = borderRadius;
-        float rsint_ = 0;
-        float dt = (0.5 * core::pi) / numCornerTriangles;
-        for (Int32 i = 1; i <= numCornerTriangles; ++i) {
-            float t = i * dt;
-            float rcost = borderRadius * std::cos(t);
-            float rsint = borderRadius * std::sin(t);
-            insertTriangle(a, r, g, b, x1_, y1_, x1_ - rcost_, y1_ - rsint_, x1_ - rcost, y1_ - rsint);
-            insertTriangle(a, r, g, b, x2_, y1_, x2_ + rsint_, y1_ - rcost_, x2_ + rsint, y1_ - rcost);
-            insertTriangle(a, r, g, b, x2_, y2_, x2_ + rcost_, y2_ + rsint_, x2_ + rcost, y2_ + rsint);
-            insertTriangle(a, r, g, b, x1_, y2_, x1_ - rsint_, y2_ + rcost_, x1_ - rsint, y2_ + rcost);
-            rcost_ = rcost;
-            rsint_ = rsint;
-        }
-    }
-}
-
-// x1, y1, x2, y2 is the text box to center the text into
-void insertText(
-        core::FloatArray& a,
-        const core::Color& c,
-        float x1, float y1, float /*x2*/, float y2,
-        const std::string& text,
-        bool hinting)
-{
-    if (text.length() > 0) {
-        float r = static_cast<float>(c[0]);
-        float g = static_cast<float>(c[1]);
-        float b = static_cast<float>(c[2]);
-
-        // Get FontFace.
-        // TODO: we should use a persistent font library rather than
-        // creating a new one each time
-        std::string facePath = core::resourcePath("graphics/fonts/SourceSansPro/TTF/SourceSansPro-Regular.ttf");
-        graphics::FontLibraryPtr fontLibrary = graphics::FontLibrary::create();
-        graphics::FontFace* fontFace = fontLibrary->addFace(facePath); // XXX can this be nullptr?
-
-        // Vertical centering
-        float ascent = static_cast<float>(fontFace->ascent());
-        float descent = static_cast<float>(fontFace->descent());
-        float height = ascent - descent;
-        float textTop = y1 + 0.5 * (y2-y1-height);
-        float baseline = textTop + ascent;
-        if (hinting) {
-            baseline = std::round(baseline);
-        }
-        core::Vec2d origin(x1, baseline);
-
-        // Shape and triangulate text
-        graphics::ShapedText shapedText = fontFace->shape(text);
-        shapedText.fill(a, origin, r, g, b);
-    }
-}
-
-core::Color getColor(const Widget* widget, core::StringId property)
-{
-    core::Color res;
-    StyleValue value = widget->style(property);
-    if (value.type() == StyleValueType::Color) {
-        res = value.color();
-    }
-    return res;
-}
-
-float getLength(const Widget* widget, core::StringId property)
-{
-    float res = 0;
-    StyleValue value = widget->style(property);
-    if (value.type() == StyleValueType::Number) {
-        res = value.number();
-    }
-    return res;
-}
-
-} // namespace
-
 void Button::onPaintDraw(graphics::Engine* engine)
 {
     if (reload_) {
         reload_ = false;
         core::FloatArray a;
-        core::Color backgroundColor = getColor(this, isHovered_ ?
+        core::Color backgroundColor = internal::getColor(this, isHovered_ ?
                         strings::background_color_on_hover :
                         strings::background_color);
-        core::Color textColor = getColor(this, strings::text_color);
-        float borderRadius = getLength(this, strings::border_radius);
+        core::Color textColor = internal::getColor(this, strings::text_color);
+        float borderRadius = internal::getLength(this, strings::border_radius);
         bool hinting = style(strings::pixel_hinting) == strings::normal;
-        insertRect(a, backgroundColor, 0, 0, width(), height(), borderRadius);
-        insertText(a, textColor, 0, 0, width(), height(), text_, hinting);
+        internal::insertRect(a, backgroundColor, 0, 0, width(), height(), borderRadius);
+        internal::insertText(a, textColor, 0, 0, width(), height(), text_, hinting);
         engine->loadTriangles(trianglesId_, a.data(), a.length());
     }
     engine->drawTriangles(trianglesId_);
