+++ libs/vgc/ui/colorpalette.cpp
@@ -22,6 +22,8 @@
 #include <vgc/core/floatarray.h>
 #include <vgc/ui/strings.h>
 
+#include <vgc/ui/internal/paintutil.h>
+
 namespace vgc {
 namespace ui {
 
@@ -123,23 +125,6 @@ void ColorPalette::onPaintCreate(graphics::Engine* engine)
 
 namespace {
 
-void insertRect(
-        core::FloatArray& a,
-        const core::Color& c,
-        float x1, float y1, float x2, float y2)
-{
-    float r = static_cast<float>(c[0]);
-    float g = static_cast<float>(c[1]);
-    float b = static_cast<float>(c[2]);
-    a.insert(a.end(), {
-        x1, y1, r, g, b,
-        x2, y1, r, g, b,
-        x1, y2, r, g, b,
-        x2, y1, r, g, b,
-        x2, y2, r, g, b,
-        x1, y2, r, g, b});
-}
-
 void insertSmoothRect(
         core::FloatArray& a,
         const core::Color& cTopLeft, const core::Color& cTopRight,
@@ -174,8 +159,8 @@ void insertCellHighlight(
 {
     // TODO: draw 4 lines (= thin rectangles) rather than 2 big rects?
     auto ch = core::Color(0.043, 0.322, 0.714); // VGC Blue
-    insertRect(a, ch, x1-2, y1-2, x2+2, y2+2);
-    insertRect(a, cellColor, x1+1, y1+1, x2-1, y2-1);
+    internal::insertRect(a, ch, x1-2, y1-2, x2+2, y2+2);
+    internal::insertRect(a, cellColor, x1+1, y1+1, x2-1, y2-1);
 }
 
 } // namespace
@@ -206,7 +191,7 @@ void ColorPalette::onPaintDraw(graphics::Engine* engine)
         float dy = std::round(dx);
         float h = startOffset + endOffset + dy * numSaturationSteps_;
         if (cellBorderWidth_ > 0 || selectorBorderWidth_ > 0) {
-            insertRect(a, borderColor_, x0, y0, x0+w, y0+h);
+            internal::insertRect(a, borderColor_, x0, y0, x0+w, y0+h);
         }
         double dhue = 360.0 / numHueSteps_;
         double hue = selectedHueIndex_ * dhue;
@@ -229,7 +214,7 @@ void ColorPalette::onPaintDraw(graphics::Engine* engine)
                 }
                 else {
                     auto c = core::Color::hsl(hue, s, l);
-                    insertRect(a, c, x1, y1, x2, y2);
+                    internal::insertRect(a, c, x1, y1, x2, y2);
                 }
             }
         }
@@ -267,7 +252,7 @@ void ColorPalette::onPaintDraw(graphics::Engine* engine)
         dy = std::round(dx);
         h = startOffset + endOffset + dy * 2;
         if (cellBorderWidth_ > 0 || selectorBorderWidth_ > 0) {
-            insertRect(a, borderColor_, x0, y0, x0+w, y0+h);
+            internal::insertRect(a, borderColor_, x0, y0, x0+w, y0+h);
         }
         double l = oldLightnessIndex_*dl;
         double s = oldSaturationIndex_*ds;
@@ -296,7 +281,7 @@ void ColorPalette::onPaintDraw(graphics::Engine* engine)
             }
             else {
                 auto c = core::Color::hsl(hue1, s, l);
-                insertRect(a, c, x1, y1, x2, y2);
+                internal::insertRect(a, c, x1, y1, x2, y2);
             }
         }
         if (isSelectedColorExact_) {
